Big CMake refactor for GLFW3 and DLL dependencies
Adding FindGLFW3 package script, adding helper IncludeGLFW3 function, changing static GLFW library to DLL version, reworking DLL export defines

+++ source/Core/Misc/Literals.cpp
@@ -4,11 +4,11 @@
 
 moe::Width_t operator "" _width(unsigned long long val)
 {
-	return moe::Width_t(val);
+	return moe::Width_t(uint32_t(val));
 }
 
 
 moe::Height_t operator "" _height(unsigned long long val)
 {
-	return moe::Height_t(val);
+	return moe::Height_t(uint32_t(val));
 }
